requirements.txt
Pillow
piexif
pypdf
python-docx
